pyproject.toml
[project]
name = "invoice-ai-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx[socks]>=0.28.1",
    "openai-agents>=0.0.14",
    "pypdf2>=3.0.1",
    "pytest>=8.3.5",
    "reportlab>=4.4.0",
    "rich>=14.0.0",
]
